requirements.txt
Flask==3.0.0
Pillow==10.2.0
vercel-wsgi==0.8.1
Werkzeug==3.0.1
